+++ src/eve-server/system/SystemManager.cpp
@@ -881,11 +881,11 @@ void SystemManager::MakeSetState(const SystemBubble *bubble, DoDestiny_SetState 
     ss.Dump( DESTINY__TRACE, "    " );
     _log( DESTINY__TRACE, "    Buffer:" );
     _hex( DESTINY__TRACE, &( ss.destiny_state->content() )[0],
-                          ss.destiny_state->content().size() );
+		(uint32)ss.destiny_state->content().size() );
 
     _log( DESTINY__TRACE, "    Decoded:" );
     Destiny::DumpUpdate( DESTINY__TRACE, &( ss.destiny_state->content() )[0],
-                                         ss.destiny_state->content().size() );
+                                         (uint32)ss.destiny_state->content().size() );
 }
 
 ItemFactory& SystemManager::itemFactory() const
